+++ packages/api-client/src/granules.ts
@@ -495,7 +495,7 @@ export const createGranule = async (params: {
  * Update/create granule in cumulus via PUT request.  Existing values will
  * be removed if not specified and in some cases replaced with defaults.
  * Granule execution association history will be retained.
- * PATCH /granules/{granuleId}
+ * PUT /granules/{collectionId}/{granuleId}
  * @param {Object} params             - params
  * @param {Object} [params.body]      - granule to pass the API lambda
  * @param {Function} params.callback  - async function to invoke the api lambda
@@ -516,7 +516,7 @@ export const replaceGranule = async (params: {
     payload: {
       httpMethod: 'PUT',
       resource: '/{proxy+}',
-      path: `/granules/${body.granuleId}`,
+      path: `/granules/${body.collectionId}/${body.granuleId}`,
       headers: {
         'Content-Type': 'application/json',
         'Cumulus-API-Version': '2',